backend/pyproject.toml
[project]
name = "beatguessr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "numpy>=2.3.3",
    "psycopg[binary]>=3.2.10",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]

[tool.black]
line-length = 120
target-version = ["py312"]
